Socket activation working with systemd
It's a bit of a hacky approach now. We start a dummy service using the
`web` name in Profile which does not listen to port 5000 that it receives.

We instead just listen on a different port (we could just not do anything as well)
and systemd starts the actual service and socket listening at port 5000.